# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Production-ready eWallet backend — auth, wallets, transactions, bills, expense tracking.",
        "title": "eWallet API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Backend Team",
            "email": "backend@ewallet.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated transaction history for the caller's wallet. Sorted by date descending (newest first). Supports filtering by type, category, and date range.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Transaction history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "send",
                            "receive",
                            "top_up",
                            "withdrawal",
                            "bill_payment"
                        ],
                        "type": "string",
                        "description": "Filter by type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "food",
                            "transport",
                            "bills",
                            "transfer",
                            "top_up",
                            "withdrawal",
                            "other"
                        ],
                        "type": "string",
                        "description": "Filter by category",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-05-01T00:00:00Z",
                        "description": "Start date — ISO 8601",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "2026-05-31T23:59:59Z",
                        "description": "End date — ISO 8601",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.PaginatedEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.TransactionResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/transactions/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transfers money from the caller's wallet to another wallet by wallet number. Both the debit and credit are atomic — if either fails the whole operation is rolled back.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Send money",
                "parameters": [
                    {
                        "description": "Transfer details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.SendRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.SendResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error / insufficient funds / self-transfer",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Recipient wallet not found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/transactions/top-up": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Adds money to the caller's own wallet. Simulates an external bank deposit (no real payment gateway in v1). Min: $1.00 (100 cents). Max: $100,000.00 (10,000,000 cents).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Top-up / Deposit",
                "parameters": [
                    {
                        "description": "Top-up details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TopUpRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TopUpResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Amount out of range",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/transactions/withdraw": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deducts money from the caller's wallet. Simulates a bank withdrawal. Min: $1.00 (100 cents).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Withdraw",
                "parameters": [
                    {
                        "description": "Withdrawal details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.WithdrawResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Insufficient funds",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    }
                }
            }
        },
        "/transactions/{transaction_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns details for one transaction. The caller must be either the sender or the recipient — returns 403 otherwise.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Get single transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction UUID",
                        "name": "transaction_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.SuccessEnvelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TransactionResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorEnvelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "details": {},
                "message": {
                    "type": "string",
                    "example": "Request validation failed."
                }
            }
        },
        "handler.ErrorEnvelope": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/handler.ErrorDetail"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "handler.PaginatedEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "type": "integer",
                    "example": 20
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "total": {
                    "type": "integer",
                    "example": 87
                },
                "total_pages": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "handler.SuccessEnvelope": {
            "type": "object",
            "properties": {
                "data": {},
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "model.SendRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "recipient_wallet_number": {
                    "type": "string"
                }
            }
        },
        "model.SendResponse": {
            "type": "object",
            "properties": {
                "new_balance": {
                    "type": "integer"
                },
                "transaction": {
                    "$ref": "#/definitions/model.TransactionResponse"
                }
            }
        },
...
        "model.TopUpRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "model.TopUpResponse": {
            "type": "object",
            "properties": {
                "new_balance": {
                    "type": "integer"
                },
                "transaction": {
                    "$ref": "#/definitions/model.TransactionResponse"
                }
            }
        },
        "model.TransactionResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "counterpart_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "note": {
                    "type": "string"
                },
                "recipient_wallet_number": {
                    "type": "string"
                },
                "sender_wallet_number": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "model.WithdrawRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "note": {
                    "type": "string"
                }
            }
        },
        "model.WithdrawResponse": {
            "type": "object",
            "properties": {
                "new_balance": {
                    "type": "integer"
                },
                "transaction": {
                    "$ref": "#/definitions/model.TransactionResponse"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your access token. Example: \"Bearer eyJhbG...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}